Refactor/iterator/live map (#8)
* Style fix

* Remove unused function

* Refactor to iterator and style fix

* Revert

* restore fill_entries()

* Maybe working

* Refactor
