+++ src/hotspot/share/gc/z/zHeap.cpp
@@ -79,79 +79,6 @@ ZHeap::ZHeap() :
   ZStatHeap::set_at_initialize(heap_min_size(), heap_max_size(), heap_max_reserve_size());
 }
 
-bool ZHeap::contains_expected(uintptr_t from) const {
-  return object_expected_dest.count(from) == 1;
-}
-
-void ZHeap::add_expected(uintptr_t from, uintptr_t to) {
-if (contains_expected(from) || contains_expected(to)) {
-    if (!(contains_expected(from) && contains_expected(to)) ||
-        !(get_expected(from) == to) ||
-        !(get_expected(to) == from)) {
-      uintptr_t f = contains_expected(from) ? get_expected(from) : 0;
-      uintptr_t t = contains_expected(to) ? get_expected(to) : 0;
-
-      std::cout << "from = " << std::hex << from << std::endl;
-      std::cout << "to = " << std::hex << to << std::endl;
-
-      std::cout << "get_expected(from) = " << std::hex << f << std::endl;
-      std::cout << "get_expected(to) = " << std::hex << t << std::endl;
-    }
-    assert(contains_expected(from) && contains_expected(to), "should already exist in both directions");
-    assert(get_expected(from) == to, "should be the same");
-    assert(get_expected(to) == from, "should be the same");
-    return;
-  }
-  object_expected_dest[from] = to;
-  object_expected_dest[to] = from;
-
-}
-
-void ZHeap::update_expected(uintptr_t from, uintptr_t to) {
-  object_expected_dest[from] = to;
-  object_expected_dest[to] = from;
-}
-
-void ZHeap::add_remap(uintptr_t from, uintptr_t to) {
-  if (contains(from) || contains(to)) {
-    if (!(contains(from) && contains(to)) ||
-        !(get_remap(from) == to) ||
-        !(get_remap(to) == from)) {
-      uintptr_t f = contains(from) ? get_remap(from) : 0;
-      uintptr_t t = contains(to) ? get_remap(to) : 0;
-
-      std::cout << "from = " << std::hex << from << std::endl;
-      std::cout << "to = " << std::hex << to << std::endl;
-
-      std::cout << "get_remap(from) = " << std::hex << f << std::endl;
-      std::cout << "get_remap(to) = " << std::hex << t << std::endl;
-    }
-    assert(contains(from) && contains(to), "should already exist in both directions");
-    assert(get_remap(from) == to, "should be the same");
-    assert(get_remap(to) == from, "should be the same");
-    return;
-  }
-  object_remaped[from] = to;
-  object_remaped[to] = from;
-}
-
-void ZHeap::remove(uintptr_t from) {
-  object_remaped.erase(from);
-}
-
-bool ZHeap::contains(uintptr_t from) const {
-  return object_remaped.count(from) == 1;
-}
-
-uintptr_t ZHeap::get_remap(uintptr_t from) const {
-  return object_remaped.at(from);
-}
-
-uintptr_t ZHeap::get_expected(uintptr_t from) const {
-  assert(contains_expected(from), "must already exist in map");
-  return object_expected_dest.at(from);
-}
-
 size_t ZHeap::heap_min_size() const {
   return MinHeapSize;
 }
@@ -499,8 +426,6 @@ void ZHeap::reset_relocation_set() {
   for (ZFragment* fragment; iter.next(&fragment);) {
     _fragment_table.remove(fragment);
   }
-  object_remaped.clear();
-  object_expected_dest.clear();
   // Reset relocation set
   _relocation_set.reset();
 }
