+++ src/hotspot/share/gc/z/zRelocate.cpp
@@ -99,7 +99,6 @@ uintptr_t ZRelocate::relocate_object_inner(ZFragment* fragment, uintptr_t from_o
   ZFragmentEntry* entry = fragment->find(from_offset);
   assert(entry != NULL, "");
   const uintptr_t to_offset = fragment->to_offset(from_offset, entry);
-  /// const uintptr_t to_offset = heap->get_expected(from_offset); //fragment->to_offset(from_offset, entry);
 
   if (entry->copied()) {
     // Already relocated, return new address
@@ -129,19 +128,11 @@ uintptr_t ZRelocate::relocate_object_inner(ZFragment* fragment, uintptr_t from_o
     uintptr_t from_offset_entry = fragment->from_offset(offset_index, (size_t)internal_index);
 
     uintptr_t to_offset = fragment->to_offset(from_offset_entry, entry);
-    /// uintptr_t to_offset = heap->get_expected(from_offset_entry);
     size_t size = ZUtils::object_size(ZAddress::good(from_offset_entry));
 
     uintptr_t from_good = ZAddress::good(from_offset_entry);
     uintptr_t to_good = ZAddress::good(to_offset);
 
-    ZHeap* heap = ZHeap::heap();
-    heap->global_lock.lock();
-    heap->add_remap(from_good, to_good);
-    heap->global_lock.unlock();
-
-
-    // assert(fragment->new_page()->is_in(to_good), "");
     ZUtils::object_copy(from_good,
                         to_good,
                         size);
@@ -159,7 +150,6 @@ uintptr_t ZRelocate::relocate_object(ZFragment* fragment, uintptr_t from_addr) c
 
   if (e->copied()) {
     uintptr_t to_good = ZAddress::good(fragment->to_offset(from_offset, e));
-    // uintptr_t to_good = ZAddress::good(heap->get_expected(from_offset)); //ZAddress::good(fragment->to_offset(from_offset, e));
     return to_good;
   }
 
@@ -168,7 +158,6 @@ uintptr_t ZRelocate::relocate_object(ZFragment* fragment, uintptr_t from_addr) c
   if (from_offset == to_offset) {
     // In-place forwarding, pin page
     assert(false, "not supported yet");
-    fragment->set_pinned();
   }
   uintptr_t to_good = ZAddress::good(to_offset);
   return to_good;
@@ -176,7 +165,6 @@ uintptr_t ZRelocate::relocate_object(ZFragment* fragment, uintptr_t from_addr) c
 
 uintptr_t ZRelocate::forward_object(ZFragment* fragment, uintptr_t from_addr) const {
   uintptr_t to_good = ZAddress::good(fragment->to_offset(ZAddress::offset(from_addr)));
-  /// uintptr_t to_good = ZAddress::good(ZHeap::heap()->get_expected(ZAddress::offset(from_addr))); //ZAddress::good(fragment->to_offset(ZAddress::offset(from_addr)));
   return to_good;
 }
 
@@ -204,13 +192,8 @@ bool ZRelocate::work(ZRelocationSetParallelIterator* iter) {
     ZRelocateObjectClosure cl(this, fragment);
     fragment->old_page()->object_iterate(&cl);
 
-    if (fragment->is_pinned()) {
-      // Relocation failed, page is now pinned
-      success = false;
-    } else {
-      // Relocation succeeded, release page
-      fragment->release_page();
-    }
+    // Relocation succeeded, release page
+    fragment->release_page();
   }
 
   return success;
